Moved existing logic to CRA
