--- old/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"><head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="stylesheet" href="./styles.css" />
-  <title>Document</title>
-</head><body>
-  <input type="checkbox" id="countryMenuVisible" />
-  <aside></aside>
-  <main>
-    <h1>
-      <div>
-        7-day moving average of COVID-19 <br />deaths per million people
-      </div>
-    </h1>
-    <div id='chart'></div>
-  </main>
-  <label for="countryMenuVisible"></label>
-  <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-  <script src="./countries.js"></script>
-  <script src="./consumer.js"></script>
-  <script src="./chart.js"></script>
-  <script src="./init.js"></script>
-</body></html>
