--- old/styles.css
@@ -1,107 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap');
-
-html, body {
-  height: 100%;
-  overflow: hidden;
-}
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: 'Poppins', sans-serif;
-}
-
-body {
-  display: flex;
-  --sidebarWidth: 300px;
-}
-
-aside {
-  width: var(--sidebarWidth);
-  height: 100%;
-  overflow-y: auto;
-}
-
-aside > label {
-  display: block;
-  padding: 10px;
-  cursor: pointer;
-}
-
-aside > label:hover {
-  background-color: hsl(0, 0%, 95%);
-}
-
-main {
-  width: calc(100% - var(--sidebarWidth));
-  height: 100%;
-}
-
-h1 {
-  height: 50px;
-  text-align: center;
-  font-size: 24px;
-  font-weight: normal;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-h1 br {
-  display: none;
-}
-
-#chart {
-  height: calc(100% - 50px);
-}
-
-#countryMenuVisible,
-label[for=countryMenuVisible] {
-  display: none;
-}
-
-@media (max-width: 1200px) {
-  h1 {
-    font-size: clamp(0px, 2.5vw, 20px);
-    line-height: 1.1;
-  }
-  h1 br {
-    display: initial;
-  }
-  #countryMenuVisible:checked + aside {
-    transform: none;
-  }
-  aside {
-    position: absolute;
-    width: 100vw;
-    background-color: white;
-    transform: translate(-100%);
-    z-index: 1;
-  }
-  main {
-    width: 100%;
-  }
-  #chart {
-    margin-right: 50px;
-  }
-  label[for=countryMenuVisible] {
-    display: block;
-    position: absolute;
-    top: 0;
-    right: 0;
-    z-index: 2;
-    width: 48px;
-    height: 48px;
-    font-size: 24px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  label[for=countryMenuVisible]:before {
-    content: '☰'
-  }
-  #countryMenuVisible:checked ~ label[for=countryMenuVisible]:before {
-    content: '❌'
-  }
-}
