+++ src/App.js
@@ -1,24 +1,103 @@
-import logo from './logo.svg';
-import './App.css';
+import countries from './countries'
+import getCountryData from './consumer'
+import drawChart from './chart'
+
+const google = window.google
+
+function debounce(functionToDebounce, wait) {
+  // From underscore.js
+	let timeout
+	return function() {
+    const context = this
+    const args = arguments
+		const later = () => {
+			timeout = null
+      functionToDebounce.apply(context, args)
+		}
+		clearTimeout(timeout)
+		timeout = setTimeout(later, wait)
+	};
+};
+
+let selectedCountries = ['denmark', 'hungary']
+
+const data = {}
+
+const init = async () => {
+  // Render sidebar
+  const sidebarHtml = countries.map(country => `
+    <label for="${country.slug}">
+      <input
+        type="checkbox"
+        id="${country.slug}"
+        ${selectedCountries.includes(country.slug) ? 'checked' : ''}
+      />
+      ${country.name}
+    </label>
+  `).join('\n')
+  document.querySelector('aside').innerHTML = sidebarHtml;
+
+  // Handle checkbox check/uncheck
+  [...document.querySelectorAll('input[type=checkbox]:not(#countryMenuVisible)')].forEach(
+    checkbox => {
+      checkbox.addEventListener('click', async event => {
+        const selectedCountry = event.target.id
+        if (selectedCountries.includes(selectedCountry)) {
+          if (selectedCountries.length <= 1) {
+            event.preventDefault()
+          }
+          else {
+            selectedCountries = selectedCountries.filter(slug => slug !== selectedCountry)
+            drawChart({data, selectedCountries})
+          }
+        }
+        else {
+          selectedCountries.push(selectedCountry)
+          if (!data[selectedCountry]) {
+            const newData = await getCountryData(selectedCountry)
+            data[selectedCountry] = newData
+          }
+          drawChart({data, selectedCountries})
+        }
+      })
+    }
+  )
+
+  // Load and show initial data
+  const initialData = await Promise.all(
+    selectedCountries.map(country => getCountryData(country))
+  )
+  selectedCountries.forEach((country, countryIndex) => {
+    data[country] = initialData[countryIndex]
+  })
+  drawChart({data, selectedCountries})
+
+  window.addEventListener('resize', debounce(function() {
+    drawChart({data, selectedCountries})
+  }, 250))
+}
+
+google.charts.load('current', {
+  packages: ['corechart']
+})
+
+google.charts.setOnLoadCallback(init)
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <>
+      <input type="checkbox" id="countryMenuVisible" />
+      <aside></aside>
+      <main>
+        <h1>
+          <div>
+            7-day moving average of COVID-19 <br />deaths per million people
+          </div>
+        </h1>
+        <div id='chart'></div>
+      </main>
+      <label htmlFor="countryMenuVisible"></label>
+    </>
   );
 }
